+++ src/test/java/TestWorkspace.java
@@ -5,10 +5,7 @@ import com.starkbank.utils.Generator;
 import org.junit.Test;
 import org.junit.Assert;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 
 public class TestWorkspace {
@@ -34,17 +31,40 @@ public class TestWorkspace {
     }
 
     @Test
-    public void testCreate() throws Exception {
-        Settings.user = utils.User.defaultOrganization();
+    public void testCreateAndUpdate() throws Exception {
+        Organization organization = utils.User.defaultOrganization();
         HashMap<String, Object> params = new HashMap<>();
         String uuid = UUID.randomUUID().toString().replace("-", "");
         params.put("username", "starkv2-" + uuid);
         params.put("name", "Stark V2: " + uuid);
-        Workspace workspace = Workspace.create(params);
+        params.put("allowedTaxIds", Arrays.asList("35953668082", "88889288043"));
+        Workspace workspace = Workspace.create(params, organization);
         Assert.assertNotNull(workspace.id);
         Assert.assertNotNull(workspace.username);
         Assert.assertNotNull(workspace.name);
         System.out.println(workspace);
+
+        String newName = "New name test";
+        String newUsername = UUID.randomUUID().toString().replace("-", "");
+        Map<String, Object> patchData = new HashMap<>();
+        List<String> allowedIds = Arrays.asList("96448045031", "26312286002");
+        patchData.put("name", newName);
+        patchData.put("username", newUsername);
+        patchData.put("allowedTaxIds", allowedIds);
+        Workspace updatedWorkspace = Workspace.update(workspace.id, patchData, Organization.replace(organization, workspace.id));
+
+        assert updatedWorkspace != null;
+        if (!updatedWorkspace.name.equals(newName)) {
+            throw new Exception("updatedWorkspace.name != " + newName);
+        }
+        if (!updatedWorkspace.username.equals(newUsername)) {
+            throw new Exception("updatedWorkspace.username != " + newUsername);
+        }
+        if (updatedWorkspace.allowedTaxIds.size() != allowedIds.size()) {
+            throw new Exception("updatedWorkspace.allowedTaxIds != " + allowedIds);
+        }
+
+        System.out.println(updatedWorkspace);
     }
 
     @Test
@@ -75,4 +95,6 @@ public class TestWorkspace {
             throw new Exception("ids.size() != 4");
         }
     }
+
+
 }